refactor: rename player to audioplayer
